+++ src/design/plone/contenttypes/restapi/services/controlpanel.py
@@ -1,15 +1,18 @@
 # -*- coding: utf-8 -*-
-from design.plone.contenttypes.controlpanels.vocabularies import (
-    IVocabulariesControlPanel,
+from design.plone.contenttypes.controlpanels.settings import (
+    IDesignPloneSettings,
+    IDesignPloneSettingsControlpanel,
 )
 from plone.restapi.controlpanels import RegistryConfigletPanel
 from zope.component import adapter
 from zope.interface import Interface
+from zope.interface import implementer
 
 
 @adapter(Interface, Interface)
-class VocabulariesControlpanel(RegistryConfigletPanel):
-    schema = IVocabulariesControlPanel
-    configlet_id = "DesignPloneVocabularies"
+@implementer(IDesignPloneSettingsControlpanel)
+class DesignPloneSettings(RegistryConfigletPanel):
+    schema = IDesignPloneSettings
+    configlet_id = "DesignPloneSettings"
     configlet_category_id = "Products"
     schema_prefix = None